ensure 8 bit
arange returns i4 for i386

+++ fitsio/hdu/table.py
@@ -1099,7 +1099,7 @@ class TableHDU(HDUBase):
             else:
                 _vstorage = vstorage
             rows = np.arange(firstrow, lastrow, step, dtype='i8')
-            sortind = np.arange(rows.size)
+            sortind = np.arange(rows.size, dtype='i8')
             colnums = self._extract_colnums()
             array = self._read_rec_with_var(
                 colnums, rows, sortind, dtype, offsets, isvar, _vstorage)
